% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Create a graph a countries population from 1990 to 2020}
\usage{
CountryPopulation(countryInput)
}
\arguments{
\item{countryInput}{A string of the country desired}
}
\value{
a graph of the country's population from 1990 to 2020.
}
\description{
Create a graph a countries population from 1990 to 2020
}
\examples{
CountryPopulation("Mexico")
CountryPopulation("Italy")
}
