% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{A series of attendence and matches of World Cups in a veriety of years and countries.}
\format{
a data frame with 22 observations
\describe{
  \item{Totalattendance}{The total attendence of the World Cup.}
  \item{Matches}{The number of matches in the World Cup.}
  \item{Averageattendence}{The average attendence of the World Cup.}
  \item{WorldCup}{Year and Country of the WorldCup.}
}
}
\source{
\url{wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
A series of attendence and matches of World Cups in a veriety of years and countries.
}
\keyword{datasets}
