% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{A time series of the population for various countries from 1950 to 2020}
\format{
a data frame with 235 observations
\describe{
  \item{Country}{The country the population was recorded for.}
  \item{1950 - 2020}{Year of the obseravtion.}
}
}
\source{
\url{population.un.org}
}
\usage{
WorldPopulation
}
\description{
A time series of the population for various countries from 1950 to 2020
}
\keyword{datasets}
